rest-api: address PR review (firmware module, in-memory download, url validation, best-effort check) + auto-init single camera and re-check FW after flash
